+++ FE/src/components/workbookPage/index.tsx
@@ -1,2 +1,5 @@
 export { default as WorkbookPageLayout } from './WorkbookPageLayout';
 export { default as WorkbookList } from './WorkbookList';
+export { default as CategoryMenu } from './CategoryMenu';
+export { default as Workbook } from './Workbook';
+export { default as GridWorkbookList } from './GridWorkbookList';
